refactor(full_statement_verifier): nest cost_model under host_utils
Same feature flag, same host-side audience; one gate instead of two.
