+++ circuit_defs/prover_examples/src/recursion.rs
@@ -210,7 +210,7 @@ mod tests {
             } else {
                 FsvProgram::UnrolledRecursionLayer
             };
-            let estimated = full_statement_verifier::cost_model::estimate_verifier_cycles(
+            let estimated = full_statement_verifier::host_utils::cost_model::estimate_verifier_cycles(
                 &proof,
                 program,
                 unrolled_blake,
