+++ gpu/program_prover/src/upstream.rs
@@ -24,7 +24,7 @@ pub use prover::definitions::SecurityLevel;
 pub use setups::UnrolledCircuitSetupParams;
 
 // The recursion protocol helpers we asked upstream to export as library code.
-pub use full_statement_verifier::cost_model::estimate_verifier_cycles;
+pub use full_statement_verifier::host_utils::cost_model::estimate_verifier_cycles;
 pub use full_statement_verifier::host_utils::{
     bridge_blake_mode, build_unified_stream, build_unrolled_stream, compute_end_params,
     final_blake_mode, load_fsv_program, unified_switch_cycles, unrolled_blake_mode,
